refactor(session-view): extract workspace-context seam
Split SessionView god seam: WorkspaceContext now owns
AsyncLocalStorage propagation (withWorkspace/workspaceCwd) and
session helpers (sessionKeyFor/execCwd/execSessionKey) with
single responsibility. SessionView keeps only served-merge
invariant (orphan healing per ADR-0008), position reconstruction
and drift computation.

Delete four compat shims (workspace, dsh-context, served-store,
drift) and update imports in tool-read/edit/undo, mutation,
read-and-serve, write-hook, hash-store, store-lifecycle,
hashline/anchor-pipeline and affected tests. Drift tests no
longer mock workspace; context tests no longer import drift.
