--- src/drift.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated — private to SessionView seam. Use `from "./session-view.js"` instead.
- * @module dsh-better-edit/drift
- */
-export * from "./session-view.js";
