--- src/served-store.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated — private to SessionView seam. Use `from "./session-view.js"` instead.
- * @module dsh-better-edit/served-store
- */
-export * from "./session-view.js";
