+++ src/tool-edit.ts
@@ -16,9 +16,9 @@ import { abortIf } from "./utils.js";
 import { execute } from "./mutation.js";
 import { EDIT_DESCRIPTION } from "./prompts.js";
 import type { FileIO } from "./fs-bridge.js";
-import { execCwd, execSessionKey } from "./session-view.js";
+import { execCwd, execSessionKey } from "./workspace-context.js";
 import type { FsSandboxController, FsEscalationArgs } from "./sandbox.js";
-import { withWorkspace } from "./session-view.js";
+import { withWorkspace } from "./workspace-context.js";
 import { findSnapshotPathsByHashes } from "./hash-store.js";
 import { parseHashRef } from "./hashline/anchor-pipeline.js";
 import type { PreparedItem } from "./edit-engine.js";
