+++ src/write-hook.ts
@@ -10,7 +10,8 @@ import type { Context } from '@deepseek-ai/cordis'
 import type { PostToolDecision, PreToolDecision, ToolExecution, ToolExecutionResult } from '@deepseek-ai/dsh-tools'
 import { readAndServe } from './read-and-serve.js'
 import type { FileIO } from './fs-bridge.js'
-import { execCwd, execSessionKey, loadServed, withWorkspace } from './session-view.js'
+import { execCwd, execSessionKey, withWorkspace } from './workspace-context.js'
+import { loadServed } from './session-view.js'
 import { HASH_SEP } from './hashline/hash-assign.js'
 import { abortIf, splitLines } from './utils.js'
