+++ test/core/workspace-store.test.ts
@@ -10,10 +10,10 @@ import { describe, expect, it } from "vitest";
 import { existsSync, mkdtempSync, rmSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { withWorkspace, workspaceCwd } from "../../src/workspace.js";
+import { withWorkspace, workspaceCwd } from "../../src/workspace-context.js";
 import { hashStorePath } from "../../src/paths.js";
 import { loadHashStore, shutdownHashStore } from "../../src/hash-store.js";
-import { recordServed, loadServed } from "../../src/served-store.js";
+import { recordServed, loadServed } from "../../src/session-view.js";
 import { lineHashes } from "../../src/hashline/index.js";
 import { initHasher } from "../../src/hashline/hasher.js";
